Move lib code into their own files
