+++ cmd/dmverity-vhd/images.go
@@ -1,0 +1,119 @@
+package main
+
+import (
+	"archive/tar"
+	"encoding/json"
+	"errors"
+	"io"
+
+	log "github.com/sirupsen/logrus"
+	"github.com/urfave/cli"
+)
+
+type LayerProcessor func(string, io.Reader) error
+
+func parseLocalImage(imageReader io.Reader, onLayer LayerProcessor) (map[int]string, map[int]string, error) {
+	log.Trace("parseLocalImage called")
+	TraceMemUsage()
+	imageFileReader := tar.NewReader(imageReader)
+	configs := make(map[string]any)
+
+	// Do a single pass of the image contents, only loading config files (not
+	// image layers) into memory. This approach is important to keep time and
+	// space complexity low when processing large images.
+	for {
+		log.Trace("looping over tar contents")
+		// Load the next file header
+		hdr, err := imageFileReader.Next()
+		if errors.Is(err, io.EOF) {
+			break
+		}
+		if err != nil {
+			return nil, nil, err
+		}
+		log.Tracef("tar hdr: %s %d", hdr.Name, hdr.Size)
+		if hdr.Typeflag != tar.TypeReg {
+			continue
+		}
+
+		imageFileReader, closer, err := decompressIfNeeded(imageFileReader)
+		imageFileReader, isTar := isTar(imageFileReader)
+		if isTar {
+			log.Infof("Found layer tarball: %s", hdr.Name)
+			if err != nil {
+				return nil, nil, err
+			}
+			if err := onLayer(hdr.Name, imageFileReader); err != nil {
+				return nil, nil, err
+			}
+			if closer != nil {
+				closer.Close()
+			}
+		} else {
+			log.Infof("Found config file: %s", hdr.Name)
+			data, err := io.ReadAll(imageFileReader)
+			if err != nil {
+				return nil, nil, err
+			}
+			var obj any
+			if err := json.Unmarshal(data, &obj); err != nil {
+				log.Tracef("Skipping non-JSON file %s: %v", hdr.Name, err)
+				continue
+			}
+			configs[hdr.Name] = obj
+		}
+	}
+
+	layerIdxToID := make(map[int]string)
+	layerIdxToPath := make(map[int]string)
+	var err error
+
+	// Different docker engine versions will either have an OCI compliant scheme
+	// for describing the image, or the older legacy docker scheme.
+	layerIdxToID, layerIdxToPath, err = parseOCIImage(configs)
+	if err == nil {
+		log.Info("OCI image format parsed successfully.")
+		return layerIdxToID, layerIdxToPath, nil
+	}
+
+	layerIdxToID, layerIdxToPath, err = parseDockerImage(configs)
+	if err == nil {
+		log.Info("Legacy docker image format parsed successfully.")
+		return layerIdxToID, layerIdxToPath, nil
+	}
+
+	// If neither format was recognized, return an error
+	return nil, nil, errors.New("image format not recognized")
+}
+
+func parseImage(ctx *cli.Context, onLayer LayerProcessor) (layerDigests map[int]string, layerIDs map[int]string, err error) {
+	imageName := ctx.String(inputFlag)
+	tarballPath := ctx.GlobalString(tarballFlag)
+	useDocker := ctx.GlobalBool(dockerFlag)
+
+	if useDocker && tarballPath != "" {
+		return nil, nil, errors.New("cannot use both docker and tarball for image source")
+	}
+
+	processLocal := func(fetcher func(string) (io.ReadCloser, error), image string) (map[int]string, map[int]string, error) {
+		imageReader, err := fetcher(image)
+		if err != nil {
+			return nil, nil, err
+		}
+		defer imageReader.Close()
+		return parseLocalImage(imageReader, onLayer)
+	}
+
+	if tarballPath != "" {
+		return processLocal(fetchImageTarball, tarballPath)
+	} else if useDocker {
+		return processLocal(fetchDockerImage, imageName)
+	} else {
+		return parseContainerRegistryImage(
+			imageName,
+			ctx.String(usernameFlag),
+			ctx.String(passwordFlag),
+			onLayer,
+		)
+	}
+}
